requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pytgcalls==3.0.0.dev24
yt-dlp==2025.1.26
python-dotenv==1.0.1
